LayoutTests/http/tests/security/contentSecurityPolicy/block-all-mixed-content/insecure-ping-in-iframe.https.html

<!DOCTYPE html>
<html>
<head>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
}
</script>
</head>
<body>
<p>This test loads a secure iframe that triggers an insecure ping load.  We should trigger a
mixed content block because the child frame has CSP directive block-all-mixed-content.</p>
<iframe id="pingFrame" src="https://127.0.0.1:8443/security/contentSecurityPolicy/block-all-mixed-content/resources/frame-with-insecure-ping.html" width="100%" height="300"></iframe>
<script>
function clickOnLinkWithPing() {
  const pingFrameDoc = pingFrame.contentDocument;
  if (window.eventSender) {
    const a = pingFrameDoc.getElementById("pingElement");
    const x = pingFrame.offsetLeft + a.offsetLeft + 2;
    const y = pingFrame.offsetTop + a.offsetTop + 2;
    eventSender.mouseMoveTo(x, y);
    eventSender.mouseDown();
    eventSender.mouseUp();
  }
}

function done()
{
    if (window.testRunner)
        testRunner.notifyDone();
}

onload = () => {
  window.setTimeout(done, 10000);
  clickOnLinkWithPing();
}
</script>
</body>
</html>
